Remove useless files and fix import errors
