+++ unbrake-frontend/src/tests/Calibration/Command.test.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Enzyme, { shallow } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import Command from "../../components/Calibration/Command";
+import Command from "../../calibration/Command";
 import ComponentTest from "../ComponentTest";
 
 Enzyme.configure({ adapter: new Adapter() });
