+++ unbrake-frontend/src/tests/Calibration/Speed.test.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Enzyme, { shallow } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import Speed from "../../components/Calibration/Speed";
+import Speed from "../../calibration/Speed";
 import ComponentTest from "../ComponentTest";
 
 Enzyme.configure({ adapter: new Adapter() });
